sdas

+++ static/js/envio.js
@@ -94,9 +94,13 @@ document.addEventListener("DOMContentLoaded", function() {
         const data = await response.json()
         
         if(data.guardado) {
+          if(confirm("¿Desea realizar la compra?")) {
             alert(data.mensaje)
-            localStorage.clear()
+            localStorage.clear()  
             location.reload()
+          } else {
+            alert('No se realizo la compra')
+          } 
         } else {
           console.log(data.mensaje)
           alert(data.mensaje)
